model_train/insightface_algorithm/requirements.txt
numpy<2
insightface
opencv-python
pillow
scikit-learn
matplotlib
onnx==1.16.1
onnxruntime
